making complementary angles labels more legible
mostly by moving them farther away from the arc. fixes #988
brainstorming here, but the problem stems from not having metrics for
the labels, so ideally we'd have some type: {width, height,
opticalCenter, trueCenter, etc} dict/obj that we could poll to safely
position a label in a tight spot. The alternative is to parametrically
position the labels further away from the angle as the angle decreases,
but this has its own set of problems.