Fix couple floating point precision issues
The root cause of exercises not accepting correct answers is due to
floating point precision on some platforms.

For issue 5198, the issue is due to the solution calculated is actually
4599.999999999. This only repros on the specific Firefox version on Mac.
The same version on Windows doesn't repro. The dividing decimals 2
exercise expects the answer to be integers so fixing the issue by
calling Math.round().

For issue 4865, the issue is likely due to the same thing. The answer
must be an integer, so also fixing the issue by calling Math.round().