New simplifying_radicals_2 exercise (with fractions)
You can now simplify fractional radicals.  Since this is all about
teaching, I'm sure it'd be better to see things like 2/2*sqrt(3/5)
instead of sqrt(3/5), so there's not that much simplification.  This is
my first exercise but it was pretty straightforward so there shouldn't
be any problems with it at all.