Remove conflicting jquery-color from utils
I noticed that the site already loads jquery-ui which provides the same
functionality as jquery-color, making it unnecessary to include it in
utils. In fact including it seemed to break the highlight effect in the
timeline when both were loaded.

This removes jquery-color from utils and the one exercise (unit_circle)
that was using it. It also adds the same version of jquery-ui used
by the site and just loads it when running in test mode.

Since future exercises will probably make use of other cool stuff in
jquery-ui it seems to make sense to do it this way.