Expose the timeline feature in test mode
With interactive answers, it's important to be able to display a
student's answer to a coach in the timeline. Previously the only way to
test the timeline was through a dev appserver instance. This adds a
"Problem history" link to debug mode so the timeline can be tested
within the standalone exercise framework.

As a bonus it adds a link to problem reports that loads the reporter's
history into the timeline viewer to see what answers they gave, if/when
they took hints, etc!