Delete Test & Fix Bugs for simplifying_radicals_2
I fixed lots of bugs and made a way for the fraction in the denominator
to be turned into a whole number.  It also explains the steps.  I also
changed some but not all fractions to \dfrac.  Bugs still in the
program are the simplification of fractions outside the radical in
certain instances (or lack thereof), and having the radical either
print a fraction when the denominator is not 1, or print the numerator
with a horizontal line under it (when it should just be a whole number).

Also deleted /test.html.