Don't disable next question button if request hasn't returned yet.
Basically this reverts most of e35e0ab3. It's acceptable if we go to the next
problem before the request has returned, which will just be another question in
the current exercise. Importantly, this gets rid of the annoying "please wait"
disable flash whenever the request is pending, and also simplifies the code.
This is also expected to be a rare scenario.