Fixed an issue where the student gets a wrong answer if they click submit before filling in the field.
The student can leave the angle at zero if they know the answer. It will not report a wrong answer in this case.

answer-types.js - For the 'sol' values that are inside a 'multiple' answer type, the attribute 'required' was added. If a required value is not set, then the student is required to fill out the value before the answer will submit. In other words, the student will not get a 'wrong' answer if they forget to fill out a value.