When asking people to determine the slope, do not show lines with slope +/-1
This is based on feedback from a user:
---
From my tutoring experience I have a suggestion for the problems
associated with 'Slope of a Line'. One class of problems gives two
points and asks the user to find the slope of the line. My suggestion
is not to include problems where the slope is 1, or -1. A common
confusion in the youth is if change in y or change in x is in the
numerator. In problems where the slope is 1, or -1 you can mix up the
dx and dy and still get the problem right leading to more confusion.
---

Slope 1/-1 is rather uncommon in this exercise anyway, so it's not a
big change to disallow it.