Try to require 'json', to get a natively-compiled json compiler.
By default, ruby uses ok_json, which is written in ruby.  The
native-compiled version should be a lot faster.  Recommendation taken
from
   http://flori.github.com/json/

If json isn't found, we'll continue, but print a warning message
saying how to make things better.  (This is in addition to the warning
message that pack.rb already printed about ok_json not being great.)