Fix #21281, #9686, #13996, #22644
When x approaches the number from the left, the denominator is negative and the hint says the lim tend -infinity.
The problem was the code didn't took in consideration the sign of the numerator, because when the numerator is negative and denominator too, the lim would tend to +infinity.
Same thing from the right side.