Add a <br /> to exploring_mean_and_median question #FIXIT
Summary:
When MathJax rendered the median at a line break, the MathJax span was
sometimes not visible. When this happens, it will magically appear as soon as
you start poking around in the DOM inspector. Rather than chasing an obscure
MathJax/browser bug, this adds a linebreak to prevent this condition from
happening in the first place.

This also changes the graphie scale to 35 from the default of 40 to fit the
narrower power-mode format on narrow screens.

Reviewers: emily

Reviewed By: emily

Differential Revision: http://phabricator.khanacademy.org/D420