Add best fit exercise
Summary:
The exercises provides a group of points, and asks whether there is a best
fit line, and if so, provide a best fit line. This uses a bunch of different
methods to check whether the line provided correctly describes a best fit
line, including the linear best fit approximation. There are two problem
types, one where the provided data make a good correlation, and one where they
do not.

Reviewers: eater

Reviewed By: eater

Differential Revision: http://phabricator.khanacademy.org/D387