Add pythagorean identity trig exercise
Summary:
Apologies in advance for the awful logic in convert-values... I considered doing something object-oriented but it
seems like it will be almost as ugly that way. Also there is some overlap in the problem types, but there is
actually significant difference in the hints for each, and since the hints are the meat of the exercise anyway,
it seemed ok to keep them as separate.

Reviewers: eater

Reviewed By: eater

CC: emily, eater

Differential Revision: http://phabricator.khanacademy.org/D601