Basic Construction Framework
Summary:
In the interest of keeping code review size down, here's the basic framework for constructions with a test exercise.
The main obstacle left is correct answer detection.

There's one UI issue that is still problematic: not sure how to deal with disposing of individual compasses/lines.
Right now you can drag compasses into a "trash" area in the bottom right, and clear the whole area, but that isn't ideal.

CC: eater

Differential Revision: http://phabricator.khanacademy.org/D667