Make answer types much more modular, and more internally consistent
Summary:
Separate answer-type creation into two steps, setup (where the answer area is
messed with) and validator creation (where a function to validate solutions is
created). These two may only communicate by a "guess" variable, generated by
the setup portion and evaluated by the validator portion. This allows for
cleaner and less hacky code, and also allows for multiple solutions to be
validated using the same setup.

Some of the khan-exercise.js code had to be modified to make this work
correctly, but most of the changes come from the rework of answer-types.js

Test Plan:
Go through lots of exercises and make sure that they still work, and that the
timeline still works with their answers

Reviewers: eater

Reviewed By: eater

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D793