added a check to see if xhr exists, that will remove the out-of-date warning on the following page load
Summary: This change will not solve people seeing the out-of-date warning message the first time, it may however fix the issue where people refresh the page and still see the out of date warning message.  Specifically if the xhr becomes undefined before this change an error was thrown and the cache never got cleared.  Deleting the fkey cookie was one way in which this seems to have happened.

Test Plan: Went to an exercise in chrome or ff.  Deleted fkey cookie. Answered a question and an error was thrown saying xhr was undefined. Because the error was thrown the Khan exercise cache did not get cleared and the red message bar did not show up.  Then reloaded the page, the wrong problem number is reloaded from the cache, did an problem, and got another page is out of date warning - this time the red warning gets called and the cache gets cleared.  Subsequent reloads and all is fine.

Reviewers: kamens

Reviewed By: kamens

Differential Revision: http://phabricator.khanacademy.org/D899