Change pack.rb to be a filter, and add a make command for old functionality.
Summary:
Now pack.rb takes an html file from stdin (or it can be listed on the
commandline, using the magic of ARGF) and emits the 'packed' html
to stdout.  This makes it easier to use in a variety of contexts.

A new 'make pack' (or 'make packed') rule mimics the old behavior:
packing everything in exercises/ and putting it in exercises-packed/
but skipping files that are newer in exercises-packed/ (and thus
unchanged since the last pack run).

Style nit: read() doesn't take parens.

Emit all errors to stderr.

Test Plan:
Ran
  mv exercises-packed /tmp
  make pack
  diff -r exercises-packed /tmp/exercises-packed
to verify that the output is unchanged.  Then ran
  make pack
again and verified it's a noop.  Then ran
  rm exercises-packed/z*
  make pack
and verified it only re-made the z_scores* packages.  Then ran
  touch -t199001010000 exercises-packed/z*
  make pack
and verified it re-made the z_scores* packages (and nothing else).

Reviewers: eater, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D995