Avoid triangles that are too small or too skinny
Fixed the issues 'many of the triangles based on independent random
points are too small or too skinny' and 'adding a constant to the side
lengths of an equilateral triangle DOES preserve similarity' using a
data-ensure to require the  triangle not be near-equilateral and its
area to be above a threshold. Also took JesseRoe's suggestion and
emphasized the words 'multiply' and 'add' with bold. Moved shared
helper functions to a script block in the header.