Domain of a function: Specify that f(x) is a real-valued function.
Summary:
If f(x) is a complex-valued function, the answers may be wrong for sqrt problem
types.

In the knowledge map, it's implicit that this exercise is from R->R because of
where it is in the knowledge map. However, when this question comes up in an
assessment, there is no such context to indicate this to the customer.

Test Plan:
locally, Chrome/OS X

Open the problem and verify the modified problem text appears for a couple of
different seeds / problem types.

Auditors: alpert, eater, jace