Turn off line-wrapping when writing .po files.
Summary:
By deafult, po-files 'wrap' each entry at 78 characters.  For the
msgid/msgstring, this is a lossless operation, but for #. comments, it
is lossy since you can't tell if a newline in a comment is a) part of
the comment-text, b) inserted due to text-wrapping, or c) inserted to
separate separate comments that were merged.

This causes problems when we insert comments with html markup.  Since
text-wrap breaks on hyphens, we could end up with text like this:
   #. You should visit <span style="left-'
   #. margin: 5px">the mall</span>.  You
   #. will like it.
which will be parsed as 'left-\nmargin', which probably won't parse.
On the other hand, we can't just remove newlines because then we'll
get 'Youwill like it.'  The only reasonable solution is to turn off
line-wrapping entirely.

Test Plan: None, but I tested the logic extensively in ka-stable, with po_sort.py

Reviewers: john

Reviewed By: john

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D2042