Remove menu and autocomplete utils/jquery-ui.js
Summary:
Neither of these appear to be needed directly in khan-exercises, and
they're both served off of the main KA environment. The duplication with
the contents here are causing issues in autocomplete behavior when
exercises are loaded.

There is still an issue of duplication for many of the remaining modules
in jquery-ui.js, such as jquery.ui.core, but that duplication doesn't
appear to be causing issues right now, and are useful for testing
khan-exercises outside of the full KA environment, so I left them in.

Test Plan:
python -m SimpleHTTPServer
navigated to a handful of exercises randomly, made sure they worked
ack '\.autocomplete\('  # no results
ack '\.menu\('  # no results

Reviewers: eater

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D2109