added skip this question A/B test into assessments
Summary: Added functionality to khan-exercise.js so that it can handle the skip question button. This goes along with review https://khanacademy.kilnhg.com/Review/K117906

Test Plan:
Load an exercise and make sure the new code doens't break anything because of the lack of a skip button
Load an assessment in a phantom window until you are in the right a/b test. Then test out clicking the skip question button multiple times quickly to make sure that only one request is sent.

Reviewers: jace, alpert, chris

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2116