Initial check-in of a new binary to decimal conversion exercise.
This exercise is designed to help people practice converting from binary to decimal.  It has 3 hints, the first hint is visual.  I'd like to work on the visual hint still, its leans some on the confusing side (perhaps its an anti-hint :), but I'm checking it in now because I think its at a reasonable baseline.

Signed-off-by: cortisclark <cortis@solrobots.com>