Correctly color history nodes for custom answer
Test plan: Answer incorrectly then correctly in an exercise with a custom answer and view the history. Looks ok. Change the exercise such that the last answer in the history is now wrong. View the same history link and see that the last answer is now red, not green.

Auditors: alpert