Improved hints based on community feedback.
I watched the Kahn Academy "Binary Numbers" video and decided to simplify the hints of this exercise based on what was taught.
- I removed hint 1 entirely as it was covered well in the video and my terse explanation, although accurate, used a lot of complicated words that could cause confusion.
- I removed exponential notation to hint 2, since the video didn't use it.
- I swapped the order of the places in hint 2 to read left to right as we are accustomed.

Problem Areas:
Unfortunately although the video is called "Binary Numbers", it never mentions the word "binary".  It also never mentions the word "decimal".  Instead it uses the terms "base 2" and "base 10", initial attempts to re-word the exercise questions and hints solely in those terms, results in awkward constructions like "What is the base 2 number 0101 in base 10?".  The exercise itself would be entitled "Convert Base 2 Numbers to Base 10 Numbers?", which doesn't seem as elegant as the current title.  The video also doesn't introduce the subscript notation for marking the type of number system used e.g.  0101₂ to indicate that the number is in binary and doesn't represent one hundred and one.

Recommendation: I recommend that a second short video be added as a supplement to the current video that names the two number systems, and introduces the subscript notation.  This will have the additional benefit of helping students to speak about binary numbers in settings outside of Kahn Academy.