Translate the Yes/No correct answer specification for Divisibility Tests. This is one grain of sand in the exercise i18n effort and is more a proof of concept than a complete solution.
Test Plan:
Go to http://localhost:8098/math/arithmetic/factors-multiples/divisibility_tests/e/divisibility_tests?lang=boxes and verify that the Yes/No options are translated into boxes.
Choose an answer at random and try to answer the problem. Do this multiple times. Verify that one of the options is treated as the correct answer for each problem, and the other
option is treated as incorrect.

Reviewers: csilvers, alpert

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2120