Replace numeric degrees with symbolic label:
Due to rounding errors when converting to niceAngles,
the numeric value displayed can lead to invalid
triangles. i.e. an SSA triangle with side, side, angle
values that produces a triangle that does not exist.

This change is to just replace the numeric rounded value
displayed with a theta character to show that the angles
are equal and that the bisector theorem can be used.