Only respect server side lastCountHints and lastAttmeptNumber if first page load.
Summary: People have said they have seen hints pre-filled in when it is not the first problem https://github.com/Khan/khan-exercises/issues/45794.  As both lastCountHints and lastAttemptNumber get set to 0 when problemDone is triggered and any out of date updateUserExercise should then get ignored in userExerciseCache _set. Hence I am not sure how this is happening, but hopefully this should fix it.  It should also make sure problems dont appear with hints filled in on the non-first problem loaded when in topic/mastery mode.

Test Plan: cross-fingers

Reviewers: alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2319

Auditors: alpert