Remove reference to non-existant "B" variable
Summary:
In the a-o-xmk and a-o-xmk2 answer-types of limits_2, there's a "B"
variable that isn't defined, but it renders fine without it (and the
question/answer still makes sense). So, if not running in debug mode,
the exception is caught and ignored, and the error goes unseen. This
removes all references to "B", which fixes everything.

Test Plan:
Visit
http://sandcastle.khanacademy.org/media/castles/xymostech:fix_limits_2/exercises/limits_2.html?debug&problem=a-o-xmk
and see that it doesn't break

Reviewers: alpert, eater

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2344