Don't allow a resubmission after q is completed
Summary:
Some students showed me a repro step of showing the red error bar that was
kind of weird. Not exactly sure how they were doing this, but if you get
a question right, and then modify the input and press enter again accidentally,
it would confuse the server and show an error bar.

This change just makes it so that you can't submit another answer once you get
it right.

Test Plan:
in an exercise, get all the hints, then get the right answer, proceeded
get a wrong answer, then get a right answer, proceeded
get a wrong answer, then a right answer, then click in the box and modify it
   and press enter and make sure it didn't do anything. click on the next
   question button and proceed

Reviewers: eater, alpert

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2356