Use plural_form to pluralize fruits/deskitems
Summary:
fruits and desk items are now Plural items, so plural_form needs to be
used to pluralize them. I'm not sure if this is the correct approach (as @john
didn't do this before) but it does make the exercise appear correctly.
Sometimes there are errors because "X" and "Y" aren't plural objects, but that
isn't shown to the user and looks fine.

Test Plan:
Go to
http://sandcastle.khanacademy.org/media/castles/xymostech:fix-linear-equations/exercises/interpreting_linear_equations.html?debug&problem=increase&seed=21
See that it doesn't show "undefined" everywhere

Reviewers: eater, alpert, john

Reviewed By: john

Differential Revision: http://phabricator.khanacademy.org/D2359