Make sure that some block elements are not in extracted strings. Add it to the linter to avoid regressions. Fix the one regression.
Summary: This makes it so that P and DIV elements contained within an extracted string and considered to be linter errors (this matches the tests in `extract_strings_test.py`).

Test Plan: Run `python lint_i18n_strings.py` and there should be no errors (although if the change to `basic_set_notation.html` is reverted then there is an error).

Reviewers: csilvers

Reviewed By: csilvers

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D2399