Get lxml to properly treat each input file as utf-8.
Summary:
lxml doesn't let you pass the encoding= argument down to html5lib, and
instead tries to guess the encoding, which depends on chardet.py being
installed on your system, which we don't do.  Sadly, it fails silently
in such cases and just assumes latin-1.

There are two reasonable ways to fix this: install chardet, or
override the lxml parser to manually set the encoding to utf-8.  I
preferred the second way, since character detection takes time and
isn't guaranteed to be accurate.  So why do it, when we know how our
files are encoded?

While cleaning up uses of lxml, got rid of one in
update_titles_tags.py that wasn't being used.

Test Plan:
Ran
   python ./clean-exercises.py ../exercises/khan-exercise.html
and saw that it did not change the input file, whereas before it would
do so in a way that munged the utf-8.

Reviewers: john, alpert

Reviewed By: john

Differential Revision: http://phabricator.khanacademy.org/D2401