Attempt to i18nize most of answer-types
Summary:
Wrap some more strings with $._, as well as try to fix up the radio
answer type to something reasonable. Before, we were directly comparing the
"None of the above" string to check if that answer was selected, which was
really ugly especially from an i18n standpoint. Now, I wrap up a lot more data
in the guess object so we can do comparisons in most places that don't involve
the raw string comparisons.

Test Plan:
In general, visit some exercises, see that they don't break. For
radio, go to
http://sandcastle.khanacademy.org/media/castles/xymostech:i18nize-answer-types/exercises/limits_1.html?debug&problem=2&seed=129
and note that the "none of the above" answer works, and that the history still
works (by clicking "problem history").

Reviewers: john, alpert, eater

Reviewed By: alpert

CC: csilvers

Differential Revision: http://phabricator.khanacademy.org/D2377