Fix up most instance of \text{} using the linter
Summary:
There are a couple cases that had to be fixed up manually, and a lot of
instances that incorrectly tried to wrap $._ around things that shouldn't be
translated (mostly the chemistry exercises). There are a couple instances that
I wasn't sure about (like the "R" in remainder) that I didn't include.

Test Plan:
Visit the changed exercises, make sure that the exercise renders and
there were no errors. Take all the hints, and make sure that instances of text
in mathjax render correctly.

Reviewers: john

Reviewed By: john

CC: alpert, eater

Differential Revision: http://phabricator.khanacademy.org/D2445