Move all files that are copies of webapp files, into a dedicated local-only dir.
Summary:
This makes it easier to keep track of these files, and also avoid
having to upload them to prod (where they're not used).

I populated the local-only dir by running the new update_local.sh
script.  I verified using 'diff' that the contents of these files is
the same before this move and after.  The one exception is
jquery.ui.*, which I've broken up into smaller files like we do in
webapp now; but I did verify that if you concatenate the files
together in the right order, you get a file that is exactly the same
as the old jquery-ui.js (excepting some file-header comments).

jquery.qtip.js has regressed between khan-exercises and webapp; here I
use the more recent of the two, which is the one in khan-exercises (I
have a commit pending in webapp to update the version there).

I incorporated the khan-exercises.css update into this new script as
well, obsoleting update-khan-site.sh.

Test Plan:
Ran
   python -mSimpleHTTPServer
in the khan-exercises rootdir, and visited
   http://localhost:8000/exercises/square_roots_2.html
and saw it render without error.  It also showed text correctly when I
hovered over 'acceptable formats', showing the qtip file is happy.

Reviewers: eater, alpert, desmond

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2493