Add a blacklist for third-party files.
Summary:
Finish the lint cleanup.

I was running runlint.py from the wrong directory, so it wasn't
linting all of khan-exercises.  Now it is.  The blacklist is a lot
bigger, and I also cleaned up files that just had a couple of lint
errors.

There should be no functional change.

Test Plan:
Ran
   ~/khan/devtools/khan-linter/runlint.py
and had it complete without error.

Ran
   khan-exercises% ~/khan/devtools/khan-linter/runlint.py
with no output.

Reviewers: alpert, eater

Reviewed By: alpert

Differential Revision: http://phabricator.khanacademy.org/D2499