Resolve some I18N todos, and simplify toSentence.
Summary:
Most of the todos I just removed as being to-done.

For toSentence(), I got rid of the 'conjunction' argument, which was
always either 'and' or the empty string.  When it was the empty
string, I just replaced it by join(', '), which is what the
exercise-author wanted to do in the first place.

Some more i18n todo-ing -- get rid of some unused functions.

Verified via
   fgrep -iw -e 'he(' -e 'an(' -e 'his(' -r exercises

Put back an() and he() -- they're used after all.

My grep was wrong.  It needs to be

   % grep -i -e '\<he(' -e '\<an(' -e '\<his(' -r exercises
   exercises/logical_arguments_deductive_reasoning.html:                        [[person(1)+" misses the bus", he(1)+" will be late for school",
   exercises/conditional_statements_2.html:                <var id="SOLUTION">["If", [an(CATEGORY), 1], ["is", 1], [an(TYPE), 1], "then", ["it", 2], [VERB_S+" "+CONCLUSION_S, 2]]</var>

But
   % grep -i '\<his(' -r exercises
still shows no hits.

Get rid of the capital-case version of some functions, which are unused.

Capitalize is ok -- javascript will do the right thing.

Fix a typo: left in toSentence().

Test Plan:
In khan-exercises root, ran
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/divisibility_intuition.html
and clicked on all the hints, until the last one said:
   The factors of 16 are 1, 2, 4, 8, and 16
which is correct serialization.  Then I visited
   http://localhost:8000/exercises/reading_bar_charts_3.html
and did problems until I got to a median problem, and clicked on all
the hints until it listed all the scores, properly separated by just
commas.  I did the exercise on www.khanacademy.org as well, to make
sure the current behavior matches.

Fingers crossed

Fingers crossed.

No results for:
   % grep -e '\<Him *(' -e '\<He *(' -e '\<An *(' -r exercises

None

Ran
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/reading_bar_charts_3.html
and did questions until I got to a 'median' question involving the
blue bars, and displayed all the hints, and saw they were rendered
properly.

Reviewers: emily, eater, john

Reviewed By: emily

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D2497