Change commafy to use icu, for the right separators for thousands and decimal.
Summary:
Add localeToFixed, which prints decimal numbers using the proper separator.

It's now a TODO to use localeToFixed() instead of toFixed(),
everywhere in khan-exercises.

Test Plan:
Ran
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/rounding_numbers.html
and did a few problems, seeing both
   Round 62.5139 to the nearest tenth.
and
   Round 46,859 to the nearest hundred.
both of which have the right separator.

Ran
   python -mSimpleHTTPServer
and visited
   http://localhost:8000/exercises/rounding_numbers.html
and clicked on the hints until they got to the number line, and saw
the number line was still showing correctly with dots.  Then
temporarily changed localeToFixed to use a comma instead of a dot, and
saw the number line had all commas.

Reviewers: emily, john, alpert

Reviewed By: alpert

CC: alpert

Maniphest Tasks: T1038

Differential Revision: http://phabricator.khanacademy.org/D2508