Move ride/rode/etc to its own exercise
Summary:
The evil ride/rode/bike/biked functions are painful for i18n, but
they're only used in a single exercise. This moves those into that exercise so
they aren't useable by the rest of the exercises.

Test Plan:
Go to `/exercises/arithmetic_word_problems_2.html?debug`, make sure
that the exercise and all the hints render correctly and that there are no
errors for all of the 9 problem types.

Reviewers: john, csilvers

Reviewed By: csilvers

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D2537