Don't use $._ inside of <var>s
Summary:
Take every use of $._ inside of normal <var>s and move them into <div
class="vars"> instead. Also, add some tests to the linter to throw errors when
$._ is used in <var>s.

Test Plan:
Run linter (not with --fix) on the exercises, see that there are no
errors related to $._ in <var>s. Visit a bunch of changed exercises locally,
make sure that there are no errors in the javascript console.

Reviewers: john, csilvers, eater

Reviewed By: csilvers

CC: alpert

Differential Revision: http://phabricator.khanacademy.org/D2628