Use Exercises instead of Khan for gotoNextExercise
This allows this even to work in perseus as well.

Test Plan:
Go to law_of_sines in local khan-exercises, see that the calculator is
cleared after a problem. Go to law_of_sines in dev appserver, see that
the calculator is cleared after a problem. Go to an exercise in perseus
that uses the calculator and see that the calculator is cleared after a
problem.

Auditors: alpert