Fix binding of newProblem event (fix matrix input)
Trello Card# 584 "Fix matrix input. It used to support arrow keys and
show brackets around the answer. What happened?"  What happened was,
newProblem used to be triggered on $(Khan), but is now triggered on
$(Exercises).  Thus, matrix input was no longer being initialized.

This problem also affected the Derivative Intuition exercise.  Points
with actual slope very close to 0 will now once again adjust to the
correct values when the problem loads.